+++ Changelog.md
@@ -4,6 +4,7 @@
 ##### 2022-XX-YY
 
 - Added link to product when listing order items
+- Added Channel to order show page (if an order has an assigned one)
 - Added Enum v4 support
 - Changed minimum AppShell requirement to v3.1.1 (to fix possible corrupted data issue with Octane)
 - Fixed Payment method listing error in case of unconfigured gateways